bot/requirements.txt
python-telegram-bot[job-queue]==21.6
aiohttp==3.10.10
python-dotenv==1.0.1
numpy==1.26.4
